# app/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server We Will Guide You server.",
        "version": "1.0.0",
        "title": "Swagger We Will Guide You",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
...
    "schemes": [
        "https"
    ],
    "paths": {
        "/manage/users": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add a new User",
                "description": "Adds a new User.",
                "operationId": "addUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "User model that needs to be added.",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/AddUserAdminApiModel"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "User added successfully",
                        "schema": {
                            "$ref": "#/definitions/DetailUserAdminApiModel"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            },
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get a list of Users",
                "description": "It's an Api for fetching All The Users.",
                "operationId": "listUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "query",
                        "name": "offset",
                        "description": "Number of entities to skip.",
                        "required": false,
                        "default": 0,
                        "type": "number"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "Number of entities to get.",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "in": "query",
                        "name": "query",
                        "description": "Query for searching and sorting.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User list",
                        "schema": {
                            "$ref": "#/definitions/UserListAdminApiModel"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            }
        },
        "/manage/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Find User by ID",
                "description": "Returns a single User model.",
                "operationId": "getUserById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of User to return.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/DetailUserAdminApiModel"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update an existing User",
                "description": "Update an existing User",
                "operationId": "updateUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "ID of updated User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Modified User model.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateUserAdminApiModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "schema": {
                            "$ref": "#/definitions/DetailUserAdminApiModel"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Deletes a User.",
                "description": "",
                "operationId": "deleteUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "User id to delete.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "User deleted successfully.",
                        "schema": {
                            "$ref": "#/definitions/VoidResponse"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            }
        },
...
        "/manage/subjects": {
            "post": {
                "tags": [
                    "Subject"
                ],
                "summary": "Add a new Subject",
                "description": "Adds a new Subject.",
                "operationId": "addSubject",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Subject model that needs to be added.",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/AddSubjectAdminApiModel"
                    }
                }],
                "responses": {
                    "201": {
                        "description": "Subject added successfully",
                        "schema": {
                            "$ref": "#/definitions/DetailSubjectGuestApiModel"
                        }
                    },
                    "409": {
                        "description": "Subject already exists",
                        "schema": {
                            "$ref": "#/definitions/VoidResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/VoidResponse"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            },
            "get": {
                "tags": [
                    "Subject"
                ],
                "summary": "Get a list of Subjects.",
                "description": "It's an Api for fetching All The Subjects.",
                "operationId": "listSubject",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "query",
                        "name": "offset",
                        "description": "Number of entities to skip.",
                        "required": false,
                        "default": 0,
                        "type": "number"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "Number of entities to get.",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "in": "query",
                        "name": "query",
                        "description": "Query for searching and sorting.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subjects found",
                        "schema": {
                            "$ref": "#/definitions/SubjectListAdminApiModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/VoidResponse"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "definitions": {
...
        "VoidResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "ResponseMessage": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "number"
                },
                "messageId": {
                    "type": "number"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "error",
                        "warning",
                        "info"
                    ]
                }
            }
        },
...
        "AddUserAdminApiModel": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "imageName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "passwordConfirmed": {
                    "type": "string"
                },
                "notificationLevel": {
                    "type": "number"
                },
                "isEnabled": {
                    "type": "boolean"
                }
            }
        },
        "UpdateUserAdminApiModel": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "imageName": {
                    "type": "string"
                },
                "notificationLevel": {
                    "type": "number"
                },
                "isEnabled": {
                    "type": "boolean"
                }
            }
        },
        "DetailUserAdminApiModel": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "firstName": {
                            "type": "string"
                        },
                        "lastName": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "phoneNumber": {
                            "type": "string"
                        },
                        "imageName": {
                            "type": "string"
                        },
                        "loginHistory": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        },
                        "notificationLevel": {
                            "type": "number"
                        },
                        "twoStepLoginType": {
                            "type": "string",
                            "enum": [
                                "email",
                                "phoneNumber",
                                "both"
                            ]
                        },
                        "isEnabled": {
                            "type": "boolean"
                        },
                        "isLockedOut": {
                            "type": "boolean"
                        },
                        "isEmailConfirmed": {
                            "type": "boolean"
                        },
                        "isPhoneNumberConfirmed": {
                            "type": "boolean"
                        },
                        "lockedOutEndDateTimeUTC": {
                            "type": "object"
                        },
                        "creationDate": {
                            "type": "object"
                        },
                        "isTwoStepAuthenticationEnabled": {
                            "type": "boolean"
                        }
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "UserListAdminApiModel": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "result": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "firstName": {
                                "type": "string"
                            },
                            "lastName": {
                                "type": "string"
                            },
                            "username": {
                                "type": "string"
                            },
                            "imageName": {
                                "type": "string"
                            },
                            "isEnabled": {
                                "type": "boolean"
                            },
                            "isLockedOut": {
                                "type": "boolean"
                            }
                        }
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "totalCount": {
                    "type": "number"
                }
            }
        },
...
        "AddSubjectAdminApiModel": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "priorityOrder": {
                    "type": "number"
                },
                "isEnabled": {
                    "type": "boolean"
                }
            }
        },
...
        "DetailSubjectGuestApiModel": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "title": {
                            "type": "string"
                        },
                        "priorityOrder": {
                            "type": "number"
                        },
                        "isEnabled": {
                            "type": "boolean"
                        }
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "SubjectListAdminApiModel": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "result": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "title": {
                                "type": "string"
                            },
                            "isEnabled": {
                                "type": "boolean"
                            }
                        }
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "totalCount": {
                    "type": "number"
                }
            }
        },
...
    },
...
}